requirements-streamlit.txt
streamlit==1.31.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
plotly==5.17.0
Flask==3.0.0
gunicorn==21.2.0
